==> chapter08/02-prep_data_for_sm_autopilot.ipynb <==
This notebook is developed using the `Python 3 (Data Science)` kernel on an `ml.t3.medium` instance.

import os
import sagemaker
sess = sagemaker.Session()
bucket = sess.default_bucket()
prefix = 'sagemaker-studio-book/chapter08'

local_prefix='winequality'
os.makedirs(local_prefix, exist_ok=True)

!wget https://archive.ics.uci.edu/ml/machine-learning-databases/wine-quality/winequality-white.csv -P {local_prefix}

import pandas as pd
from sklearn.model_selection import train_test_split
df = pd.read_csv(f'{local_prefix}/winequality-white.csv', delimiter = ';')

df.head()

df_train, df_test = train_test_split(df, test_size = 0.1, random_state = 42)

df_train.shape, df_test.shape

df_train['quality'].value_counts()

df_train.to_csv(f'{local_prefix}/winequality-white-train.csv', index=False)
df_test.to_csv(f'{local_prefix}/winequality-white-test.csv', index=False)
df_test.drop(columns=['quality']).to_csv(f'{local_prefix}/winequality-white-test-notarget.csv', 
                                         index=False, header=False)

sagemaker.s3.S3Uploader.upload(local_path=f'{local_prefix}/winequality-white-train.csv',
                               desired_s3_uri=f's3://{bucket}/{prefix}/{local_prefix}',
                               sagemaker_session=sess)

sagemaker.s3.S3Uploader.upload(local_path=f'{local_prefix}/winequality-white-test.csv',
                               desired_s3_uri=f's3://{bucket}/{prefix}/{local_prefix}',
                               sagemaker_session=sess)

df_test.drop(columns=['quality']).to_csv(f'{local_prefix}/winequality-white-test-notarget.csv', 
                                         index=False, header=False)
sagemaker.s3.S3Uploader.upload(local_path=f'{local_prefix}/winequality-white-test-notarget.csv',
                               desired_s3_uri=f's3://{bucket}/{prefix}/{local_prefix}',
                               sagemaker_session=sess)
